+++ pkg/providers/appsec/provider.go
@@ -180,7 +180,9 @@ func (p *Subprovider) FrameworkResources() []func() resource.Resource {
 // FrameworkDataSources returns the appsec data sources implemented using terraform-plugin-framework
 func (p *Subprovider) FrameworkDataSources() []func() datasource.DataSource {
 	return []func() datasource.DataSource{
-		NewRapidRulesDataSource,
 		NewCustomRulesUsageDataSource,
+		NewRapidRulesDataSource,
+		NewURLProtectionRuleActionsDataSource,
+		NewURLProtectionRulesActionsDataSource,
 	}
 }