+++ pkg/providers/appsec/testdata/TestDSURLProtectionRulesActions/missing_security_policy_id.tf
@@ -1,0 +1,7 @@
+provider "akamai" {
+  edgerc = "~/.edgerc"
+}
+
+data "akamai_appsec_url_protection_rules_actions" "test" {
+  config_id = 43253
+}