+++ CHANGELOG.md
@@ -115,6 +115,8 @@
 
 
 
+* Appsec
+ * Added datasources to support URL Protection rule actions `akamai_appsec_url_protection_rules_actions` and `akamai_appsec_url_protection_rule_actions`.
 
 
 * PAPI